backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
langchain>=0.2.6
langchain-google-genai>=1.0.6
python-dotenv>=1.0.1
pydantic>=2.7.4
beautifulsoup4>=4.12.3
requests>=2.32.3
